requirements.txt
google-genai
python-dotenv
gspread
pandas
streamlit>=1.30.0
altair<5
extra-streamlit-components
gallery-dl
